feat: add coordinate related functions.
